Store all data points in the database and as files
Until now, everything, with some exceptions such as band structures,
has been stored in the database. This includes the main part of the
data sets, which are the numerical values. Now, the data points are
stored in both the database and on the hard drive in the form of
files. This way, the original content as uploaded by the user is saved
without modifications.

This is especially beneficial for atomic structures, where multiple
storage formats can now be supported. At the moment, only the FHI-aims
and CIF formats are recognized, but more could easily be added.