Allow user to upload geometry in any format
If the geometry format is not recognized, the lattice constants and
angles are not automatically populated. Once those are manually
filled, the data can be uploaded and viewed in JSMol (assuming it is
viewable in JSMol).

Fix minor bugs related to importing lattice parameter data from files
for multiple subsets.