Save original import file name
Should the user click on the "Import from file" button to autofill the
textarea, use that file name for saving the data on the hard
drive. Previously, the data file was always called data1.txt
(data2.txt, ...). Now, that will be the fallback if the original file
name is not known (if the user does not autofill the textarea but
inserts the values manually).