Show atomic structure on the data set page
That is, the JSmol visualization is shown on the page of each data
set, not just on the system page (assuming the primary property is the
atomic structure). When viewing all atomic structures for a given
system, however, the structure is shown only for the first data
set. When trying to show it for the other data sets, JSmol hangs for
some reason.