Update base.html
This updates the documentation link for the database software (MatD3), not for the database itelf.  The update is needed since the previous readthedocs setup was no longer supported by readthedocs due to a "secret" that is now considered deprecated.

The actual HybriD3 database documentation link at https://hybrid3.readthedocs.io/en/latest will likely require a similar update - not yet tested. This will need to be done later.